break out generator_ids_cache into a library called async-cache
