+++ src/rust/async-cache/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "async-cache"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+futures = "0.3"
+moka = { version = "0.9", features = ["future"] }
+thiserror = "1.0"
+tokio = { version = "1.17", features = ["macros", "rt", "rt-multi-thread"] }
+tracing = "0.1"
